Update GitHub action versions in deploy.yaml
The commit updates the action versions used in deploy.yaml workflow. The actions/checkout, docker/build-push-action, actions/upload-artifact, actions/download-artifact, and docker/login-action have all been updated to their respective newer versions. This aims to improve workflow efficiency, and keep the setup up-to-date with the latest features and bug fixes of the tools.